made tons of style changes
