+++ package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "babel-jest": "^24.9.0",
     "cypress": "^4.7.0",
+    "cypress-cucumber-preprocessor": "^2.3.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-config-airbnb-base": "^14.1.0",
